Added spacial hashing FluidContainer. Fixed issues with new kernels

+++ assignment2/FluidSimulation/FluidSimulation/Particle.h
@@ -20,7 +20,7 @@ public:
 	float getWGrad(float r);
 	float getWLaplacian(float r);
 	float getWVisLaplacian(float r);
-	float getWPoly6(float r);
+	double getWPoly6(double r);
 	Vec2f getWSpikyGrad(Vec2f r);
 
 	int m_index;
